Added function to get the number of users connected to the server over json.
The new methods are:

Send "getusersonline" as method returns "usersonline" with "data" set to the number of clients connected to the server.

Made functions for usersinroom and getusersinroom, but this needs a little creative thinking first.
"getusersinroom", can be called, but it will return 0 until the function is created to find the number of users in a perticular room.